+++ src/app/(providers)/(styles)/my-challenge/_components/RoutineCheckBox/RoutineCheckBox.tsx
@@ -10,7 +10,9 @@ import { v4 } from "uuid"
 
 import { RoutineDoneType } from "../../../../../../../types/routineDone.type"
 import { RoutineType } from "../../../../../../../types/supabase.type"
-import { MyChallengePageContext } from "../../context"
+import useMyChallengePageContext, {
+  MyChallengePageContext,
+} from "../../context"
 
 interface RoutineCheckBoxProps {
   milestoneId: string
@@ -31,7 +33,7 @@ function RoutineCheckBox({
   routineDoneDailyId,
 }: PropsWithChildren<RoutineCheckBoxProps>) {
   const routineCount = routines.length
-  const { todayDate } = useContext(MyChallengePageContext)
+  const { todayDate } = useMyChallengePageContext()
   // 전체 routine_done에서 routine_done_daily_id를 통해서
   // 현재 체크한 루틴에 대한 오늘 날짜의 데이터 가져오기 및 존재 여부 확인하며,
   // 이를 활용해 추후 체크박스의 최초값(선택/해제) 부여
